+++ python/numpyqi/random.py
@@ -98,7 +98,8 @@ def rand_choi_op(dim_in, dim_out, seed=None):
     np0 = scipy.linalg.expm(tmp0 + tmp0.T.conj())
     tmp1 = np.einsum(np0.reshape(dim_in, dim_out, dim_in, dim_out), [0,1,2,1], [0,2], optimize=True)
     # matrix square root should be the same
-    tmp2 = np.linalg.inv(scipy.linalg.sqrtm(tmp1))
+    tmp2 = np.linalg.inv(scipy.linalg.sqrtm(tmp1).astype(tmp1.dtype))
+    # TODO .astype(xxx.dtype) scipy-v1.10 bug https://github.com/scipy/scipy/issues/18250
     # tmp2 = np.linalg.inv(scipy.linalg.cholesky(tmp1))
     ret = np.einsum(np0.reshape(dim_in,dim_out,dim_in,dim_out), [0,1,2,3],
             tmp2.conj(), [0,4], tmp2, [2,5], [4,1,5,3], optimize=True).reshape(N0,N0)
